Split Codex repository evidence parser
