+++ crates/ctx-history-capture/src/provider/codex/repository.rs
@@ -1,7 +1,4 @@
-use std::{
-    collections::{HashMap, HashSet},
-    path::Path,
-};
+use std::{collections::HashSet, path::Path};
 
 use ctx_history_core::{RepositoryFileObservationKind, CORE_REPOSITORY_OUTCOME_CAPTURE_REVISION};
 use serde_json::{json, Map, Value};
@@ -15,18 +12,16 @@ use crate::{
 
 #[path = "repository/outcomes.rs"]
 mod outcomes;
+mod static_js;
 
 pub(crate) use outcomes::CodexRepositoryResultEvidence;
+use static_js::{StaticJsParser, StaticNestedToolCall};
 
 const MAX_STRUCTURED_ARGUMENT_BYTES: usize = 16 * 1024 * 1024;
 const MAX_WORKDIR_BYTES: usize = 16 * 1024;
 const MAX_CALL_ID_BYTES: usize = 1024;
 const MAX_CONTINUATION_CELL_ID_BYTES: usize = 1024;
-const MAX_STATIC_NESTED_TOOL_CALLS: usize = 24;
-const MAX_STATIC_BINDINGS: usize = 24;
 const MAX_STATIC_PATCH_PATHS: usize = 256;
-const MAX_STATIC_LITERAL_DEPTH: usize = 32;
-const MAX_STATIC_LITERAL_ITEMS: usize = 256;
 const CODEX_CONTINUATION_CALL_ID_DOMAIN: &[u8] = b"ctx/codex-nativepath/continuation-call-id/v1\0";
 const CODEX_COMMAND_DOMAIN: &[u8] = b"ctx/codex-nativepath/exact-command/v1\0";
 
@@ -444,458 +439,6 @@ fn bounded_patch_path(path: &str) -> Option<String> {
     bounded_path(path).then(|| path.to_owned())
 }
 
-enum StaticNestedToolCall {
-    ExecCommand(Map<String, Value>),
-    ApplyPatch(String),
-}
-
-struct StaticJsParser<'a> {
-    source: &'a [u8],
-    cursor: usize,
-    static_strings: HashMap<String, String>,
-}
-
-impl<'a> StaticJsParser<'a> {
-    fn new(source: &'a str) -> Self {
-        Self {
-            source: source.as_bytes(),
-            cursor: 0,
-            static_strings: HashMap::new(),
-        }
-    }
-
-    fn parse_program(mut self) -> Option<Vec<StaticNestedToolCall>> {
-        let mut calls = Vec::new();
-        let mut terminal_output_statements = 0_usize;
-        loop {
-            self.skip_program_trivia()?;
-            if self.cursor == self.source.len() {
-                break;
-            }
-            if terminal_output_statements == 0 {
-                if let Some((name, value)) = self.parse_static_string_declaration() {
-                    if self.static_strings.len() >= MAX_STATIC_BINDINGS
-                        || self.static_strings.insert(name, value).is_some()
-                    {
-                        return None;
-                    }
-                    continue;
-                }
-                if let Some(call) = self.parse_tool_statement() {
-                    calls.push(call);
-                    if calls.len() > MAX_STATIC_NESTED_TOOL_CALLS {
-                        return None;
-                    }
-                    continue;
-                }
-                if let Some(call) = self.parse_wrapped_tool_statement() {
-                    calls.push(call);
-                    if calls.len() > MAX_STATIC_NESTED_TOOL_CALLS {
-                        return None;
-                    }
-                    continue;
-                }
-            }
-            if self.parse_terminal_output_statement() {
-                terminal_output_statements += 1;
-                if terminal_output_statements > MAX_STATIC_NESTED_TOOL_CALLS {
-                    return None;
-                }
-                continue;
-            }
-            return None;
-        }
-        Some(calls)
-    }
-
-    fn parse_static_string_declaration(&mut self) -> Option<(String, String)> {
-        let checkpoint = self.cursor;
-        let parsed = self.parse_static_string_declaration_inner();
-        if parsed.is_none() {
-            self.cursor = checkpoint;
-        }
-        parsed
-    }
-
-    fn parse_static_string_declaration_inner(&mut self) -> Option<(String, String)> {
-        self.consume_keyword("const").then_some(())?;
-        self.skip_whitespace();
-        let name = self.parse_identifier()?;
-        self.skip_whitespace();
-        self.consume_byte(b'=').then_some(())?;
-        self.skip_whitespace();
-        let value = self.parse_json_string()?;
-        self.consume_statement_terminator().then_some(())?;
-        Some((name, value))
-    }
-
-    fn parse_terminal_output_statement(&mut self) -> bool {
-        let checkpoint = self.cursor;
-        if !self.consume_keyword("text") {
-            return false;
-        }
-        self.skip_whitespace();
-        if !self.consume_byte(b'(') {
-            self.cursor = checkpoint;
-            return false;
-        }
-        self.skip_whitespace();
-        let argument_ok = if self.source.get(self.cursor) == Some(&b'`') {
-            self.parse_output_template()
-        } else if self.source.get(self.cursor) == Some(&b'"') {
-            self.parse_json_string().is_some()
-        } else {
-            self.parse_member_reference()
-        };
-        if !argument_ok {
-            self.cursor = checkpoint;
-            return false;
-        }
-        self.skip_whitespace();
-        if !self.consume_byte(b')') {
-            self.cursor = checkpoint;
-            return false;
-        }
-        if !self.consume_statement_terminator() {
-            self.cursor = checkpoint;
-            return false;
-        }
-        true
-    }
-
-    fn parse_output_template(&mut self) -> bool {
-        if !self.consume_byte(b'`') {
-            return false;
-        }
-        while let Some(byte) = self.source.get(self.cursor).copied() {
-            match byte {
-                b'`' => {
-                    self.cursor += 1;
-                    return true;
-                }
-                b'\\' => {
-                    self.cursor += 1;
-                    if self.source.get(self.cursor).is_none() {
-                        return false;
-                    }
-                    self.cursor += 1;
-                }
-                b'$' if self
-                    .source
-                    .get(self.cursor.saturating_add(1))
-                    .is_some_and(|next| *next == b'{') =>
-                {
-                    self.cursor += 2;
-                    self.skip_whitespace();
-                    if !self.parse_member_reference() {
-                        return false;
-                    }
-                    self.skip_whitespace();
-                    if !self.consume_byte(b'}') {
-                        return false;
-                    }
-                }
-                _ => self.cursor += 1,
-            }
-        }
-        false
-    }
-
-    fn parse_member_reference(&mut self) -> bool {
-        if self.parse_identifier().is_none() {
-            return false;
-        }
-        loop {
-            if !self.consume_byte(b'.') {
-                return true;
-            }
-            if self.parse_identifier().is_none() {
-                return false;
-            }
-        }
-    }
-
-    fn consume_statement_terminator(&mut self) -> bool {
-        self.skip_whitespace();
-        if self.consume_byte(b';') {
-            return true;
-        }
-        let saved = self.cursor;
-        if self.skip_program_trivia().is_some() && self.cursor == self.source.len() {
-            self.cursor = saved;
-            true
-        } else {
-            self.cursor = saved;
-            false
-        }
-    }
-
-    fn parse_tool_statement(&mut self) -> Option<StaticNestedToolCall> {
-        let checkpoint = self.cursor;
-        let parsed = self.parse_tool_statement_inner();
-        if parsed.is_none() {
-            self.cursor = checkpoint;
-        }
-        parsed
-    }
-
-    fn parse_tool_statement_inner(&mut self) -> Option<StaticNestedToolCall> {
-        if self.consume_keyword("const") {
-            self.skip_whitespace();
-            self.parse_identifier()?;
-            self.skip_whitespace();
-            self.consume_byte(b'=').then_some(())?;
-            self.skip_whitespace();
-        }
-        let call = self.parse_tool_invocation()?;
-        self.consume_statement_terminator().then_some(())?;
-        Some(call)
-    }
-
-    fn parse_wrapped_tool_statement(&mut self) -> Option<StaticNestedToolCall> {
-        let checkpoint = self.cursor;
-        let parsed = self.parse_wrapped_tool_statement_inner();
-        if parsed.is_none() {
-            self.cursor = checkpoint;
-        }
-        parsed
-    }
-
-    fn parse_wrapped_tool_statement_inner(&mut self) -> Option<StaticNestedToolCall> {
-        self.consume_keyword("text").then_some(())?;
-        self.skip_whitespace();
-        self.consume_byte(b'(').then_some(())?;
-        self.skip_whitespace();
-        let call = self.parse_tool_invocation()?;
-        self.skip_whitespace();
-        self.consume_byte(b')').then_some(())?;
-        self.consume_statement_terminator().then_some(())?;
-        Some(call)
-    }
-
-    fn parse_tool_invocation(&mut self) -> Option<StaticNestedToolCall> {
-        self.consume_keyword("await").then_some(())?;
-        self.skip_whitespace();
-        self.consume_bytes(b"tools.").then_some(())?;
-        let method = self.parse_identifier()?;
-        self.skip_whitespace();
-        self.consume_byte(b'(').then_some(())?;
-        self.skip_whitespace();
-        let value = if method == "apply_patch"
-            && self
-                .source
-                .get(self.cursor)
-                .is_some_and(|byte| byte.is_ascii_alphabetic() || matches!(byte, b'_' | b'$'))
-        {
-            let binding = self.parse_identifier()?;
-            Value::String(self.static_strings.get(&binding)?.clone())
-        } else {
-            self.parse_static_value(0)?
-        };
-        self.skip_whitespace();
-        self.consume_byte(b')').then_some(())?;
-        match (method.as_str(), value) {
-            ("exec_command", Value::Object(arguments)) => {
-                Some(StaticNestedToolCall::ExecCommand(arguments))
-            }
-            ("apply_patch", Value::String(patch)) => Some(StaticNestedToolCall::ApplyPatch(patch)),
-            _ => None,
-        }
-    }
-
-    fn parse_static_value(&mut self, depth: usize) -> Option<Value> {
-        if depth > MAX_STATIC_LITERAL_DEPTH {
-            return None;
-        }
-        self.skip_whitespace();
-        match self.source.get(self.cursor).copied()? {
-            b'"' => self.parse_json_string().map(Value::String),
-            b'{' => self.parse_static_object(depth + 1).map(Value::Object),
-            b'[' => self.parse_static_array(depth + 1).map(Value::Array),
-            b't' if self.consume_keyword("true") => Some(Value::Bool(true)),
-            b'f' if self.consume_keyword("false") => Some(Value::Bool(false)),
-            b'n' if self.consume_keyword("null") => Some(Value::Null),
-            b'-' | b'0'..=b'9' => self.parse_number(),
-            _ => None,
-        }
-    }
-
-    fn parse_static_object(&mut self, depth: usize) -> Option<Map<String, Value>> {
-        self.consume_byte(b'{').then_some(())?;
-        let mut object = Map::new();
-        loop {
-            self.skip_whitespace();
-            if self.consume_byte(b'}') {
-                return Some(object);
-            }
-            if object.len() >= MAX_STATIC_LITERAL_ITEMS {
-                return None;
-            }
-            let key = if self.source.get(self.cursor) == Some(&b'"') {
-                self.parse_json_string()?
-            } else {
-                self.parse_identifier()?
-            };
-            self.skip_whitespace();
-            self.consume_byte(b':').then_some(())?;
-            let value = self.parse_static_value(depth)?;
-            if object.insert(key, value).is_some() {
-                return None;
-            }
-            self.skip_whitespace();
-            if self.consume_byte(b'}') {
-                return Some(object);
-            }
-            self.consume_byte(b',').then_some(())?;
-        }
-    }
-
-    fn parse_static_array(&mut self, depth: usize) -> Option<Vec<Value>> {
-        self.consume_byte(b'[').then_some(())?;
-        let mut values = Vec::new();
-        loop {
-            self.skip_whitespace();
-            if self.consume_byte(b']') {
-                return Some(values);
-            }
-            if values.len() >= MAX_STATIC_LITERAL_ITEMS {
-                return None;
-            }
-            values.push(self.parse_static_value(depth)?);
-            self.skip_whitespace();
-            if self.consume_byte(b']') {
-                return Some(values);
-            }
-            self.consume_byte(b',').then_some(())?;
-        }
-    }
-
-    fn parse_json_string(&mut self) -> Option<String> {
-        let start = self.cursor;
-        self.consume_byte(b'"').then_some(())?;
-        let mut escaped = false;
-        while let Some(byte) = self.source.get(self.cursor).copied() {
-            self.cursor += 1;
-            if escaped {
-                escaped = false;
-            } else if byte == b'\\' {
-                escaped = true;
-            } else if byte == b'"' {
-                return serde_json::from_slice(self.source.get(start..self.cursor)?).ok();
-            } else if byte.is_ascii_control() {
-                return None;
-            }
-        }
-        None
-    }
-
-    fn parse_number(&mut self) -> Option<Value> {
-        let start = self.cursor;
-        while self.source.get(self.cursor).is_some_and(|byte| {
-            byte.is_ascii_digit() || matches!(byte, b'-' | b'+' | b'.' | b'e' | b'E')
-        }) {
-            self.cursor += 1;
-        }
-        serde_json::from_slice(self.source.get(start..self.cursor)?).ok()
-    }
-
-    fn parse_identifier(&mut self) -> Option<String> {
-        let start = self.cursor;
-        let first = self.source.get(self.cursor).copied()?;
-        if !first.is_ascii_alphabetic() && !matches!(first, b'_' | b'$') {
-            return None;
-        }
-        self.cursor += 1;
-        while self
-            .source
-            .get(self.cursor)
-            .is_some_and(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'_' | b'$'))
-        {
-            self.cursor += 1;
-        }
-        std::str::from_utf8(self.source.get(start..self.cursor)?)
-            .ok()
-            .map(str::to_owned)
-    }
-
-    fn consume_keyword(&mut self, keyword: &str) -> bool {
-        let start = self.cursor;
-        if !self.consume_bytes(keyword.as_bytes()) {
-            return false;
-        }
-        if self
-            .source
-            .get(self.cursor)
-            .is_some_and(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'_' | b'$'))
-        {
-            self.cursor = start;
-            return false;
-        }
-        true
-    }
-
-    fn consume_bytes(&mut self, expected: &[u8]) -> bool {
-        if self
-            .source
-            .get(self.cursor..self.cursor.saturating_add(expected.len()))
-            == Some(expected)
-        {
-            self.cursor += expected.len();
-            true
-        } else {
-            false
-        }
-    }
-
-    fn consume_byte(&mut self, expected: u8) -> bool {
-        if self.source.get(self.cursor) == Some(&expected) {
-            self.cursor += 1;
-            true
-        } else {
-            false
-        }
-    }
-
-    fn skip_whitespace(&mut self) {
-        while self
-            .source
-            .get(self.cursor)
-            .is_some_and(|byte| byte.is_ascii_whitespace())
-        {
-            self.cursor += 1;
-        }
-    }
-
-    fn skip_program_trivia(&mut self) -> Option<()> {
-        loop {
-            self.skip_whitespace();
-            if self.source.get(self.cursor..self.cursor.saturating_add(2)) == Some(b"//") {
-                self.cursor += 2;
-                while self
-                    .source
-                    .get(self.cursor)
-                    .is_some_and(|byte| *byte != b'\n')
-                {
-                    self.cursor += 1;
-                }
-                continue;
-            }
-            if self.source.get(self.cursor..self.cursor.saturating_add(2)) == Some(b"/*") {
-                self.cursor += 2;
-                while self.source.get(self.cursor..self.cursor.saturating_add(2)) != Some(b"*/") {
-                    self.cursor += 1;
-                    if self.cursor >= self.source.len() {
-                        return None;
-                    }
-                }
-                self.cursor += 2;
-                continue;
-            }
-            return Some(());
-        }
-    }
-}
-
 fn bounded_absolute_path(value: &str) -> bool {
     bounded_path(value) && Path::new(value).is_absolute()
 }
@@ -999,340 +542,4 @@ fn hex_digest(value: &[u8; 32]) -> String {
 }
 
 #[cfg(test)]
-mod tests {
-    use ctx_history_core::{RepositoryAbstentionReason, RepositoryFileObservationKind};
-    use serde_json::json;
-
-    use super::repository_tool_evidence;
-    use crate::provider::codex::events::CodexToolCallContext;
-    use crate::repository_attribution::{attribute, AttributionInput, CommandEvidenceDisposition};
-    use crate::{OutputOutcome, OutputOutcomeMetadata};
-
-    #[test]
-    fn accepts_only_one_top_level_native_argument_decode_and_redacts_it() {
-        let payload = json!({
-            "type": "function_call",
-            "name": "exec_command",
-            "call_id": "call-1",
-            "arguments": json!({
-                "cmd": "git status",
-                "workdir": "/repo",
-                "yield_time_ms": 10000,
-                "decoy": {"cmd": "git commit -m decoy", "workdir": "/other"}
-            }).to_string()
-        });
-        let evidence = repository_tool_evidence(&payload).remove(0);
-        assert_eq!(evidence.command.as_deref(), Some("git status"));
-        assert_eq!(evidence.declared_workdir.as_deref(), Some("/repo"));
-        let encoded = serde_json::to_string(&evidence.structured_content).unwrap();
-        assert!(!encoded.contains("git status"));
-        assert!(!encoded.contains("decoy"));
-        assert_eq!(
-            evidence.structured_content["provider_native_tool"]["raw_arguments_retained"],
-            false
-        );
-    }
-
-    #[test]
-    fn oversized_native_command_retains_typed_abstention_and_blocks_cwd_fallback() {
-        let temp = tempfile::tempdir().unwrap();
-        assert!(std::process::Command::new("/usr/bin/git")
-            .args(["init", "-q"])
-            .current_dir(temp.path())
-            .status()
-            .unwrap()
-            .success());
-        let oversized = "x".repeat(super::MAX_COMMAND_BYTES + 1);
-        let payload = json!({
-            "type": "function_call",
-            "name": "exec_command",
-            "call_id": "oversized-command",
-            "arguments": json!({"cmd": oversized}).to_string(),
-        });
-        let evidence = repository_tool_evidence(&payload).remove(0);
-        assert!(evidence.command.is_none());
-        assert!(evidence.command_too_large);
-
-        let annotation = attribute(AttributionInput {
-            session_cwd: Some(temp.path().to_string_lossy().into_owned()),
-            command: evidence.command,
-            command_disposition: CommandEvidenceDisposition::CommandTooLarge,
-            ..AttributionInput::default()
-        });
-        assert!(annotation.repository_bindings.is_empty());
-        assert!(annotation.repository_abstentions.iter().any(|abstention| {
-            abstention.reason == RepositoryAbstentionReason::CommandTooLarge
-        }));
-    }
-
-    #[test]
-    fn javascript_wrappers_nested_only_commands_and_missing_call_ids_abstain() {
-        for payload in [
-            json!({
-                "type": "custom_tool_call",
-                "name": "exec_command",
-                "call_id": "call-1",
-                "arguments": "tools.exec_command({cmd:'git status',workdir:'/repo'})"
-            }),
-            json!({
-                "type": "function_call",
-                "name": "exec_command",
-                "call_id": "call-2",
-                "arguments": {"dead_branch": {"cmd": "git status", "workdir": "/repo"}}
-            }),
-            json!({
-                "type": "function_call",
-                "name": "exec_command",
-                "arguments": {"cmd": "git status", "workdir": "/repo"}
-            }),
-            json!({
-                "type": "function_call",
-                "name": "exec_command",
-                "call_id": "call-3",
-                "arguments": {"cmd": "git status"},
-                "input": {"cmd": "git commit -m decoy"}
-            }),
-            json!({
-                "type": "function_call",
-                "name": "exec_command",
-                "tool": "wait",
-                "call_id": "call-4",
-                "arguments": {"cmd": "git status"}
-            }),
-        ] {
-            assert!(repository_tool_evidence(&payload).is_empty());
-        }
-    }
-
-    #[test]
-    fn exact_top_level_literal_calls_decode_commands_and_patch_headers() {
-        let payload = json!({
-            "type": "custom_tool_call",
-            "name": "exec",
-            "call_id": "outer-1",
-            "input": r#"
-                const first = await tools.exec_command({cmd:"git status",workdir:"/repo/one",yield_time_ms:10000});
-                const second = await tools.exec_command({cmd:"git log -1",workdir:"/repo/two"});
-                const patched = await tools.apply_patch("*** Begin Patch\n*** Add File: /repo/one/src/new.rs\n*** Update File: /repo/one/src/lib.rs\n*** Delete File: /repo/one/src/old.rs\n*** End Patch");
-                text(first.output);
-            "#
-        });
-        let evidence = repository_tool_evidence(&payload);
-        assert_eq!(evidence.len(), 3);
-        assert_eq!(evidence[0].tool_name, "exec_command");
-        assert_eq!(evidence[0].declared_workdir.as_deref(), Some("/repo/one"));
-        assert_eq!(evidence[0].command.as_deref(), Some("git status"));
-        assert_eq!(evidence[1].declared_workdir.as_deref(), Some("/repo/two"));
-        assert_eq!(evidence[1].command.as_deref(), Some("git log -1"));
-        assert_eq!(evidence[2].tool_name, "apply_patch");
-        assert_eq!(evidence[2].file_observations.len(), 3);
-        assert_eq!(
-            evidence[2].file_observations[0].path,
-            "/repo/one/src/new.rs"
-        );
-        assert_eq!(
-            evidence[2].file_observations[0].kind,
-            RepositoryFileObservationKind::Created
-        );
-        assert_eq!(
-            evidence[2].file_observations[1].path,
-            "/repo/one/src/lib.rs"
-        );
-        assert_eq!(
-            evidence[2].file_observations[1].kind,
-            RepositoryFileObservationKind::Modified
-        );
-        assert_eq!(
-            evidence[2].file_observations[2].path,
-            "/repo/one/src/old.rs"
-        );
-        assert_eq!(
-            evidence[2].file_observations[2].kind,
-            RepositoryFileObservationKind::Deleted
-        );
-    }
-
-    #[test]
-    fn genuine_terminal_template_preserves_declared_workdir_for_call_and_linked_result() {
-        let temp = tempfile::tempdir().unwrap();
-        assert!(std::process::Command::new("/usr/bin/git")
-            .args(["init", "-q"])
-            .current_dir(temp.path())
-            .status()
-            .unwrap()
-            .success());
-        let workdir = temp.path().to_string_lossy();
-        let payload = json!({
-            "type": "custom_tool_call",
-            "name": "exec",
-            "call_id": "outer-template",
-            "input": format!(
-                "const r = await tools.exec_command({{\"cmd\":\"git diff --check && git diff && cargo test\",\"workdir\":{workdir:?},\"yield_time_ms\":30000}});\ntext(r.output);\ntext(`exit=${{r.exit_code}}`);\n"
-            ),
-        });
-        let mut evidence = repository_tool_evidence(&payload);
-        assert_eq!(evidence.len(), 1);
-        let evidence = evidence.remove(0);
-        assert_eq!(evidence.tool_name, "exec_command");
-        assert_eq!(evidence.declared_workdir.as_deref(), Some(workdir.as_ref()));
-        assert_eq!(
-            evidence.command.as_deref(),
-            Some("git diff --check && git diff && cargo test")
-        );
-
-        let call_annotation = attribute(AttributionInput {
-            declared_tool_workdir: evidence.declared_workdir.clone(),
-            command: evidence.command.clone(),
-            ..AttributionInput::default()
-        });
-        assert_eq!(call_annotation.repository_bindings.len(), 1);
-
-        let context = CodexToolCallContext {
-            tool_name: evidence.tool_name,
-            exact_command: evidence.command,
-            declared_workdir: evidence.declared_workdir,
-            origin_call_id: Some("outer-template".to_owned()),
-            origin_event_sequence: Some(39),
-            ..CodexToolCallContext::default()
-        };
-        let result = super::repository_result_evidence(
-            &json!({"output": "Script completed\nWall time 0.1 seconds\nOutput:\n"}),
-            &context,
-            "outer-template",
-            [7; 32],
-            10,
-            &OutputOutcomeMetadata {
-                outcome: OutputOutcome::Success,
-                exit_code: Some(0),
-                duration_ms: Some(100),
-            },
-        )
-        .unwrap();
-        assert_eq!(result.declared_workdir.as_deref(), Some(workdir.as_ref()));
-        let result_annotation = attribute(AttributionInput {
-            declared_tool_workdir: result.declared_workdir,
-            command: result.command,
-            outcome_operation_repository_path: result.outcome_operation_repository_path,
-            outcome_output_repository_path: result.outcome_output_repository_path,
-            outcome_observations: result.outcomes,
-            outcome_abstentions: result.abstentions,
-            ..AttributionInput::default()
-        });
-        assert_eq!(result_annotation.repository_bindings.len(), 1);
-        assert_eq!(
-            result_annotation.repository_bindings[0].binding_id,
-            call_annotation.repository_bindings[0].binding_id
-        );
-    }
-
-    #[test]
-    fn genuine_bound_patch_wrapper_emits_exact_file_observations() {
-        let payload = json!({
-            "type": "custom_tool_call",
-            "name": "exec",
-            "call_id": "outer-patch",
-            "input": r#"
-                const patch = "*** Begin Patch\n*** Add File: /repo/src/new.rs\n*** Update File: /repo/src/lib.rs\n*** Move to: /repo/src/moved.rs\n*** Delete File: /repo/src/old.rs\n*** End Patch";
-                text(await tools.apply_patch(patch));
-            "#,
-        });
-        let evidence = repository_tool_evidence(&payload);
-        assert_eq!(evidence.len(), 1);
-        assert_eq!(evidence[0].tool_name, "apply_patch");
-        assert_eq!(evidence[0].file_observations.len(), 3);
-        assert_eq!(evidence[0].file_observations[0].path, "/repo/src/new.rs");
-        assert_eq!(
-            evidence[0].file_observations[0].kind,
-            RepositoryFileObservationKind::Created
-        );
-        assert_eq!(evidence[0].file_observations[1].path, "/repo/src/moved.rs");
-        assert_eq!(
-            evidence[0].file_observations[1].prior_path.as_deref(),
-            Some("/repo/src/lib.rs")
-        );
-        assert_eq!(
-            evidence[0].file_observations[1].kind,
-            RepositoryFileObservationKind::Renamed
-        );
-        assert_eq!(evidence[0].file_observations[2].path, "/repo/src/old.rs");
-        assert_eq!(
-            evidence[0].file_observations[2].kind,
-            RepositoryFileObservationKind::Deleted
-        );
-        assert_eq!(
-            evidence[0].structured_content["provider_native_tool"]["argument_schema"],
-            "codex_nested_apply_patch_literal_v3"
-        );
-    }
-
-    #[test]
-    fn direct_native_patch_input_uses_the_same_bounded_parser() {
-        let payload = json!({
-            "type": "custom_tool_call",
-            "name": "apply_patch",
-            "call_id": "direct-patch",
-            "input": "*** Begin Patch\n*** Update File: src/lib.rs\n@@\n-old\n+new\n*** End Patch\n",
-        });
-        let evidence = repository_tool_evidence(&payload);
-        assert_eq!(evidence.len(), 1);
-        assert_eq!(evidence[0].tool_name, "apply_patch");
-        assert_eq!(evidence[0].file_observations.len(), 1);
-        assert_eq!(evidence[0].file_observations[0].path, "src/lib.rs");
-        assert_eq!(
-            evidence[0].file_observations[0].kind,
-            RepositoryFileObservationKind::Modified
-        );
-    }
-
-    #[test]
-    fn inert_or_dynamic_javascript_never_emits_executed_tool_evidence() {
-        for source in [
-            r#"const example = "tools.exec_command({cmd:\"git commit -m inert\",workdir:\"/repo\"})"; text(example);"#,
-            r#"// tools.exec_command({cmd:"git commit -m comment",workdir:"/repo"})
-                text("*** Add File: /repo/inert.rs");"#,
-            r#"if (false) { await tools.exec_command({cmd:"git commit -m dead",workdir:"/repo"}); }"#,
-            r#"const args = {cmd:"git commit -m dynamic",workdir:"/repo"};
-                const result = await tools.exec_command(args);"#,
-            r#"const result = await tools.exec_command({cmd:"git status",workdir:"/repo"});
-                observeDynamically(result);"#,
-            r#"text(prior.output);
-                const result = await tools.exec_command({cmd:"git status",workdir:"/repo"});"#,
-            r#"const patch = "*** Begin Patch\n*** Add File: /repo/inert.rs\n*** End Patch"; text(patch);"#,
-            r#"const patch = choosePatch(); text(await tools.apply_patch(patch));"#,
-            r#"const patch = "*** Begin Patch\n*** Add File: /repo/dynamic.rs\n*** End Patch";
-                text(await tools.apply_patch(transform(patch)));"#,
-            r#"const result = await tools.exec_command({cmd:"git status",workdir:"/repo"});
-                text(`${sideEffect()}`);"#,
-        ] {
-            let payload = json!({
-                "type": "custom_tool_call",
-                "name": "exec",
-                "call_id": "outer-inert",
-                "input": source,
-            });
-            assert!(repository_tool_evidence(&payload).is_empty(), "{source}");
-        }
-    }
-
-    #[test]
-    fn continuation_controls_require_exact_bounded_identifiers() {
-        assert_eq!(
-            super::running_continuation_cell_id(&json!({
-                "output": "Script running with cell ID cell-7\n"
-            }))
-            .as_deref(),
-            Some("cell-7")
-        );
-        assert!(super::running_continuation_cell_id(&json!({
-            "output": "prose says Script running with cell ID cell-7"
-        }))
-        .is_none());
-        assert!(super::terminal_continuation_result(&json!({
-            "output": "Script completed\nFinal output:\nok"
-        })));
-        assert!(!super::terminal_continuation_result(&json!({
-            "output": "Script completed",
-            "result": "Process exited with code 0"
-        })));
-    }
-}
+mod tests;
